GCW-3462 Create dom test utils for login page
